# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "系统后端 API",
        "title": "Go DDD Seed API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://github.com/go-ddd-seed/go-ddd-seed/support",
            "email": "team@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/permissions": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "查询所有权限列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "权限管理-权限"
                ],
                "summary": "权限列表",
                "responses": {
                    "200": {
                        "description": "权限列表",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_go-ddd-seed_go-ddd-seed_internal_trigger_http_resp.PermissionResp"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "创建新的权限定义",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "权限管理-权限"
                ],
                "summary": "创建权限",
                "parameters": [
                    {
                        "description": "创建权限请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_go-ddd-seed_go-ddd-seed_internal_trigger_http_req.CreatePermissionReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_go-ddd-seed_go-ddd-seed_internal_trigger_http_resp.PermissionResp"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/docs.APIError"
                        }
                    }
                }
            }
        },
        "/permissions/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "更新指定权限的名称和描述",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "权限管理-权限"
                ],
                "summary": "更新权限",
                "parameters": [
                    {
                        "type": "string",
                        "description": "权限ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新权限请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_go-ddd-seed_go-ddd-seed_internal_trigger_http_req.UpdatePermissionReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_go-ddd-seed_go-ddd-seed_internal_trigger_http_resp.PermissionResp"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/docs.APIError"
                        }
                    },
                    "404": {
                        "description": "权限不存在",
                        "schema": {
                            "$ref": "#/definitions/docs.APIError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "删除指定的权限",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "权限管理-权限"
                ],
                "summary": "删除权限",
                "parameters": [
                    {
                        "type": "string",
                        "description": "权限ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/docs.APIError"
                        }
                    },
                    "404": {
                        "description": "权限不存在",
                        "schema": {
                            "$ref": "#/definitions/docs.APIError"
                        }
                    }
                }
            }
        },
        "/roles": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "查询所有角色列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "权限管理-角色"
                ],
                "summary": "角色列表",
                "responses": {
                    "200": {
                        "description": "角色列表",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_go-ddd-seed_go-ddd-seed_internal_trigger_http_resp.RoleResp"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "创建新的角色定义",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "权限管理-角色"
                ],
                "summary": "创建角色",
                "parameters": [
                    {
                        "description": "创建角色请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_go-ddd-seed_go-ddd-seed_internal_trigger_http_req.CreateRoleReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_go-ddd-seed_go-ddd-seed_internal_trigger_http_resp.RoleResp"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/docs.APIError"
                        }
                    }
                }
            }
        },
        "/roles/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据ID获取角色详情",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "权限管理-角色"
                ],
                "summary": "按ID查询角色",
                "parameters": [
                    {
                        "type": "string",
                        "description": "角色ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "角色详情",
                        "schema": {
                            "$ref": "#/definitions/github_com_go-ddd-seed_go-ddd-seed_internal_trigger_http_resp.RoleResp"
                        }
                    },
                    "404": {
                        "description": "角色不存在",
                        "schema": {
                            "$ref": "#/definitions/docs.APIError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "更新指定角色的名称和描述",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "权限管理-角色"
                ],
                "summary": "更新角色",
                "parameters": [
                    {
                        "type": "string",
                        "description": "角色ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新角色请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_go-ddd-seed_go-ddd-seed_internal_trigger_http_req.UpdateRoleReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_go-ddd-seed_go-ddd-seed_internal_trigger_http_resp.RoleResp"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/docs.APIError"
                        }
                    },
                    "404": {
                        "description": "角色不存在",
                        "schema": {
                            "$ref": "#/definitions/docs.APIError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "删除指定的角色",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "权限管理-角色"
                ],
                "summary": "删除角色",
                "parameters": [
                    {
                        "type": "string",
                        "description": "角色ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/docs.APIError"
                        }
                    },
                    "404": {
                        "description": "角色不存在",
                        "schema": {
                            "$ref": "#/definitions/docs.APIError"
                        }
                    }
                }
            }
        },
        "/roles/{id}/disable": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "禁用指定的角色",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "权限管理-角色"
                ],
                "summary": "禁用角色",
                "parameters": [
                    {
                        "type": "string",
                        "description": "角色ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "禁用成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_go-ddd-seed_go-ddd-seed_internal_trigger_http_resp.RoleResp"
                        }
                    },
                    "404": {
                        "description": "角色不存在",
                        "schema": {
                            "$ref": "#/definitions/docs.APIError"
                        }
                    }
                }
            }
        },
        "/roles/{id}/enable": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "启用指定的角色",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "权限管理-角色"
                ],
                "summary": "启用角色",
                "parameters": [
                    {
                        "type": "string",
                        "description": "角色ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "启用成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_go-ddd-seed_go-ddd-seed_internal_trigger_http_resp.RoleResp"
                        }
                    },
                    "404": {
                        "description": "角色不存在",
                        "schema": {
                            "$ref": "#/definitions/docs.APIError"
                        }
                    }
                }
            }
        },
        "/roles/{id}/permissions": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "为指定角色批量分配权限",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "权限管理-角色"
                ],
                "summary": "为角色分配权限",
                "parameters": [
                    {
                        "type": "string",
                        "description": "角色ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "分配权限请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_go-ddd-seed_go-ddd-seed_internal_trigger_http_req.AssignPermissionReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "分配成功",
                        "schema": {
                            "$ref": "#/definitions/docs.APIError"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/docs.APIError"
                        }
                    },
                    "404": {
                        "description": "角色或权限不存在",
                        "schema": {
                            "$ref": "#/definitions/docs.APIError"
                        }
                    }
                }
            }
        },
        "/roles/{id}/permissions/{permId}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "从指定角色中移除特定权限",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "权限管理-角色"
                ],
                "summary": "移除角色权限",
                "parameters": [
                    {
                        "type": "string",
                        "description": "角色ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "权限ID",
                        "name": "permId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "移除成功",
                        "schema": {
                            "$ref": "#/definitions/docs.APIError"
                        }
                    },
                    "404": {
                        "description": "角色或权限不存在",
                        "schema": {
                            "$ref": "#/definitions/docs.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "docs.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 2
                },
                "message": {
                    "type": "string",
                    "example": "请求参数错误"
                }
            }
        },
...
        "github_com_go-ddd-seed_go-ddd-seed_internal_trigger_http_resp.PermissionResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Bearer token authentication. Format: \"Bearer \u003ctoken\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}